refactor: Move, document `NarwhalsAggregation`
